refactor: move internal/install to internal/local
