+++ internal/local/bundle/manifest.go
