+++ internal/local/k3s/images.go
@@ -10,7 +10,7 @@ import (
 	"os"
 	"runtime"
 
-	"github.com/weka/gohomecli/internal/install/bundle"
+	"github.com/weka/gohomecli/internal/local/bundle"
 	"github.com/weka/gohomecli/internal/utils"
 )
