+++ internal/local/k3s/install.go
@@ -12,7 +12,7 @@ import (
 	"regexp"
 	"strings"
 
-	"github.com/weka/gohomecli/internal/install/bundle"
+	"github.com/weka/gohomecli/internal/local/bundle"
 	"github.com/weka/gohomecli/internal/utils"
 )
