+++ internal/local/web/api/shared.go
