+++ internal/local/web/frontend/public/logo512.png
