+++ internal/local/web/frontend/public/manifest.json
